refactor(tui): extract render layer to app/render.rs (S4 PR 5) — clears ADR ceiling
Fifth and final S4 slice for the app.rs ceiling. The per-screen renderers
(render_overview, render_chat_summary, render_memory_summary,
render_sessions_summary, render_vault_summary, render_cases_summary,
render_system_summary) plus the line-pushing helpers (append_transcript_entry,
append_provider_status_rows, append_model_rows, telegram_status_lines,
telegram_send_record_lines, provider_rows_as_lines) move into app/render.rs.

Mechanics:
  - new file: crates/memphis-tui/src/app/render.rs (~615 LoC)
  - app/mod.rs: 2359 → 1771 LoC (-588)
  - 11 entry points pub(super); 2 internal helpers private
  - imports tightened: removed now-unused ChatTranscriptEntry from
    mod.rs's memphis_operator import

🎯 ADR-001 ceiling cleared:
  - app/mod.rs:          1771 LoC  (was 6039)
  - app/host_results.rs: 1323 LoC
  - app/tests.rs:        1576 LoC
  - app/commands.rs:      671 LoC
  - app/render.rs:        615 LoC
  - app/format.rs:        196 LoC
  All under the 2000 LoC limit for crates/.

Rubric impact: criterion #4 (file-size violations) goes 1 → 0.

LoC trajectory across S4:
  app.rs (start)       6039  ← ADR violation (>2000)
  → after PR 1         5395  (-644 → commands.rs)
  → after PR 2         4100  (-1295 → host_results.rs)
  → after PR 3         2530  (-1570 → tests.rs)
  → after PR 4         2359  (-171 → format.rs)
  → after PR 5         1771  (-588 → render.rs) ✓

84/84 TUI tests pass throughout. No behaviour change.
